dashboard: added subject template
