+++ packages/core/mailer/minibars.test.js
@@ -1158,22 +1158,27 @@ const test_set = [
                                     <p class="f-fallback" style="font-size: 12px; line-height: 1.625; color: #85878E; margin: 0;">Amount</p>
                                   </th>
                                 </tr>
+                                
                                 <tr>
                                   <td width="80%" class="purchase_item" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 15px; color: #51545E; line-height: 18px; padding: 10px 0;"><span class="f-fallback">Robot Leg White</span></td>
                                   <td class="align-right" width="20%" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 16px; text-align: right;" align="right"><span class="f-fallback">3</span></td>
                                 </tr>
+                                
                                 <tr>
                                   <td width="80%" class="purchase_item" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 15px; color: #51545E; line-height: 18px; padding: 10px 0;"><span class="f-fallback">Battery</span></td>
                                   <td class="align-right" width="20%" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 16px; text-align: right;" align="right"><span class="f-fallback">5</span></td>
                                 </tr>
+                                
                                 <tr>
                                   <td width="80%" class="purchase_item" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 15px; color: #51545E; line-height: 18px; padding: 10px 0;"><span class="f-fallback">Robot Arm Red</span></td>
                                   <td class="align-right" width="20%" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 16px; text-align: right;" align="right"><span class="f-fallback">2</span></td>
                                 </tr>
+                                
                                 <tr>
                                   <td width="80%" class="purchase_item" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 15px; color: #51545E; line-height: 18px; padding: 10px 0;"><span class="f-fallback">Robot Arm Green</span></td>
                                   <td class="align-right" width="20%" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 16px; text-align: right;" align="right"><span class="f-fallback">1</span></td>
                                 </tr>
+                                
                                 <tr>
                                   <td width="80%" class="purchase_footer table-border-color" valign="middle" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 16px; padding-top: 15px; border-top-width: 1px; border-top-style: solid;">
                                     <p class="f-fallback purchase_total purchase_total--label" style="font-size: 14px; line-height: 1.625; text-align: right; font-weight: bold; color: #333333; margin: 0; padding: 0 15px 0 0;" align="right">Shipping</p>
@@ -2496,6 +2501,7 @@ const test_set = [
                         <table class="body-action" align="center" width="100%" cellpadding="0" cellspacing="0" role="presentation" style="width: 100%; -premailer-width: 100%; -premailer-cellpadding: 0; -premailer-cellspacing: 0; text-align: center; margin: 0px auto; padding: 0;">
                           <tr>
                             <td align="center" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 16px;">
+                              
                               <table width="100%" border="0" cellspacing="0" cellpadding="50" role="presentation">
                                 <tr>
                                   <td align="center" style="word-break: break-word; font-family: &quot;Nunito Sans&quot;, Helvetica, Arial, sans-serif; font-size: 16px;">
@@ -2505,6 +2511,7 @@ const test_set = [
                                   </td>
                                 </tr>
                               </table>
+                              
 
 
                             </td>
@@ -2551,21 +2558,30 @@ const test_set = [
         "action_title": "Take Action"
       }
     }
+  },
+
+
+  { // test empty and missing data compilation
+    handlebars_template: `Hello {{name}}, happy to see you`,
+    handlebars_expected: `Hello , happy to see you`,
+    input: {}
   }
 ]
+
 
 test(
   'compare minibars to handlebars',
   async () => {
-    for(const item of test_set.slice(0,1)) {
+    for(const item of test_set.slice(0)) {
       const result = Minibars.compile(
         item.handlebars_template,
       )(item.input);
-      
-      // assert.snapshot(
-      //   result,
-      //   item.handlebars_expected
-      // )
+      // console.log(result);
+      // return;
+      assert.equal(
+        result,
+        item.handlebars_expected
+      )
     }
     
   }
