+++ packages/dashboard/src/comps/template-example-input.jsx
@@ -1,199 +1,6 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { Editor, useMonaco} from "@monaco-editor/react";
 import useDarkMode from '@/hooks/use-dark-mode.js'
-import CapsulesView from './capsules-view.jsx';
-
-
-const info_example = {
-  general_store_name: 'Wush Wush Games',
-  general_store_website: 'https://wush.games/',
-  general_store_logo_url: undefined,
-  general_store_description: 'We sell retro video games',
-  general_confirm_email_base_url: 'https://wush.games/confirm-email',
-  general_store_support_email: 'support@wush.games'
-}
-
-/**
- * @type {import('@storecraft/core/api').OrderData}
- */
-const example_order = {
-  contact: {
-    email: 'john@doe.com',
-    firstname: 'John',
-    phone_number: '000-000-000',
-    customer_id: 'cus_65f2ae6e8bf30e6cd0ca95fa'
-  },
-  address: {
-
-  },
-  "status": {
-    "checkout": {
-      "id": 0,
-      "name2": "created",
-      "name": "Created"
-    },
-    "payment": {
-      "id": 1,
-      "name": "Authorized",
-      "name2": "authorized"
-    },
-    "fulfillment": {
-      "id": 0,
-      "name2": "draft",
-      "name": "Draft"
-    }
-  },
-  "pricing": {
-    evo: [
-      {
-        quantity_discounted: 0,
-        quantity_undiscounted: 11,
-        subtotal: 1100,
-        total: 1150
-      },
-      {
-        quantity_discounted: 2,
-        total_discount: 100,
-        quantity_undiscounted: 9,
-        discount: {
-          "active": true,
-          "handle": "discount-bundle-50-off-robot-arms-and-legs-not-recursive",
-          "title": "50% OFF Bundle: robot arms and legs (not recursive)",
-          "priority": 0,
-          "application": {
-            "id": 0,
-            "name": "Automatic",
-            "name2": "automatic"
-          },
-          "info": {
-            "details": {
-              "meta": {
-                "id": 4,
-                "type": "bundle",
-                "name": "Bundle Discount"
-              },
-              "extra": {
-                "fixed": 0,
-                "percent": 50,
-                "recursive": false
-              }
-            },
-            "filters": [
-              {
-                "meta": {
-                  "id": 4,
-                  "type": "product",
-                  "op": "p-in-tags",
-                  "name": "Product has Tag"
-                },
-                "value": [
-                  "robot_arm"
-                ]
-              },
-              {
-                "meta": {
-                  "id": 4,
-                  "type": "product",
-                  "op": "p-in-tags",
-                  "name": "Product has Tag"
-                },
-                "value": [
-                  "robot_leg"
-                ]
-              }
-            ]
-          }
-        },
-        discount_code: 'discount-bundle-50-off-robot-arms-and-legs-not-recursive',
-        subtotal: 1000,
-        total: 1050
-      }
-    ],
-    uid: undefined,
-    shipping_method: { title: '', handle: '', price: 50 },
-    subtotal_discount: 100,
-    subtotal_undiscounted: 1100,
-    subtotal: 1000,
-    total: 1050,
-    quantity_total: 11,
-    quantity_discounted: 2,
-    errors: []
-  },
-  "line_items": [
-    {
-      id: 'robot-leg-white', qty: 3, 
-      data: { 
-        tags: ['robot_leg'], 
-        qty: 100, 
-        active: true, title: 'Robot Leg White', 
-        price: 100 
-      }
-    },
-    {
-      id: 'battery', qty: 5, 
-      data: { 
-        tags: ['would-not-be-discounted'], 
-        qty: 100, 
-        active: true, title: 'Battery', 
-        price: 100 
-      }
-    },
-    {
-      id: 'robot-arm-red', qty: 2, 
-      data: { 
-        tags: ['robot_arm'], 
-        qty: 100, 
-        active: true, title: 'Robot Arm Red', 
-        price: 100 
-      }
-    },
-    {
-      id: 'robot-arm-green', qty: 1, 
-      data: { 
-        tags: ['robot_arm'], 
-        qty: 100, 
-        active: true, title: 'Robot Arm Green', 
-        price: 100 
-      }
-    },
-  ],
-  "shipping_method": {
-    "handle": "ship-fast",
-    "title": "ship fast",
-    "price": 50
-  },
-  "id": "order_65d774c6445e4581b9e34c11",
-  "created_at": "2024-02-22T16:22:30.095Z",
-  "updated_at": "2024-02-22T16:22:30.095Z"
-}
-
-/**
- * @type {import('@storecraft/core/api').CustomerType}
- */
-const example_customer = {
-  "email": "john@dow.com",
-  "firstname": "John",
-  "lastname": "Dow",
-  "id": "cus_65f2ae6e8bf30e6cd0ca95fa",
-}
-
-
-const capsules = [
-  {
-    name: 'customer',
-    example: JSON.stringify({
-      customer: example_customer,
-      info: info_example
-    })
-  },
-  {
-    name: 'order',
-    example: JSON.stringify({
-      order: example_order,
-      info: info_example
-    })
-  }
-]
 
 
 /**
@@ -511,26 +318,18 @@ export const TemplateExampleInput = (
 <div {...rest} >
   <div className='flex flex-col w-full gap-5'>
 
-    <div className='flex flex-row flex-wrap gap-3'>
-      <CapsulesView 
-          tags={capsules} 
-          name_fn={it => it.name} 
-          onClick={it => { editor_onChange(it.example, null) }} />
-
-    </div>
-  
     <Editor
-        options={{tabSize: 2}}
-        width='100%'
-        height="200px"
-        className='rounded-md border shelf-border-color overflow-clip'
-        onChange={editor_onChange}
-        value={source}
-        theme={darkMode ? 'coblat' : 'light'}
-        defaultLanguage='json'
-        // defaultValue="// some comment"
-        // onMount={handleEditorDidMount}
-      />
+      options={{tabSize: 2, minimap: { enabled: false }}}
+      width='100%'
+      height="300px"
+      className='rounded-md border shelf-border-color overflow-clip'
+      onChange={editor_onChange}
+      value={source}
+      theme={darkMode ? 'coblat' : 'light'}
+      defaultLanguage='json'
+      // defaultValue="// some comment"
+      // onMount={handleEditorDidMount}
+    />
 
   </div>
 </div>
